drop v2 as we pre-v1
