--- .github/workflows/ci-v2.yml
@@ -1,46 +1,0 @@
-name: build-v2
-
-on:
-  push:
-    branches:
-    tags:
-    paths:
-      - ".github/workflows/ci-v2.yml"
-      - "v2/**"
-  pull_request:
-    paths:
-      - ".github/workflows/ci-v2.yml"
-      - "v2/**"
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: set up go
-        uses: actions/setup-go@v5
-        with:
-          go-version: "1.24"
-        id: go
-
-      - name: checkout
-        uses: actions/checkout@v4
-
-      - name: build and test
-        run: |
-          go test -timeout=60s -v -race -p 1 -covermode=atomic -coverprofile=$GITHUB_WORKSPACE/profile.cov ./...
-          go build -race
-        working-directory: v2
-
-      - name: golangci-lint
-        uses: golangci/golangci-lint-action@v4
-        with:
-          version: latest
-          working-directory: v2
-
-      - name: submit coverage
-        run: |
-          go install github.com/mattn/goveralls@latest
-          goveralls -service="github" -coverprofile=$GITHUB_WORKSPACE/profile.cov
-        env:
-          COVERALLS_TOKEN: ${{ secrets.GITHUB_TOKEN }}
