+++ .github/workflows/ci.yml
@@ -11,10 +11,10 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: set up go 1.20
+      - name: set up go 1.24
         uses: actions/setup-go@v3
         with:
-          go-version: "1.20"
+          go-version: "1.24"
         id: go
 
       - name: checkout
@@ -26,12 +26,9 @@ jobs:
           go test -timeout=60s -race -covermode=atomic -coverprofile=$GITHUB_WORKSPACE/profile.cov_tmp
           cat $GITHUB_WORKSPACE/profile.cov_tmp | grep -v "_mock.go" > $GITHUB_WORKSPACE/profile.cov
           go build -race
-        env:
-          GO111MODULE: "on"
-          TZ: "America/Chicago"
 
       - name: golangci-lint
-        uses: golangci/golangci-lint-action@v3
+        uses: golangci/golangci-lint-action@v4
         with:
           version: latest
