+++ .golangci.yml
@@ -1,14 +1,10 @@
 linters-settings:
   govet:
-    check-shadowing: true
-  golint:
-    min-confidence: 0.6
+    shadow: true
   gocyclo:
     min-complexity: 15
   maligned:
     suggest-new: true
-  dupl:
-    threshold: 100
   goconst:
     min-len: 2
     min-occurrences: 2
@@ -23,61 +19,60 @@ linters-settings:
       - experimental
     disabled-checks:
       - wrapperFunc
-      - hugeParam
-      - rangeValCopy
 
 linters:
-  disable-all: true
   enable:
-    - megacheck
+    - staticcheck
     - revive
     - govet
     - unconvert
-    - gas
-    - misspell
-    - unused
+    - gosec
+    - unparam
     - typecheck
     - ineffassign
     - stylecheck
     - gochecknoinits
-    - exportloopref
+    - copyloopvar
+    - gocritic
     - nakedret
     - gosimple
     - prealloc
-
+    - unused
+    - contextcheck
+    - copyloopvar
+    - decorder
+    - errorlint
+    - exptostd
+    - gochecknoglobals
+    - gofmt
+    - goimports
+    - nilerr
+    - predeclared
+    - testifylint
+    - thelper
   fast: false
+  disable-all: true
 
 
 run:
-  #  modules-download-mode: vendor
-  skip-dirs:
-    - vendor
   concurrency: 4
 
 issues:
   exclude-rules:
-    - text: "should have a package comment, unless it's in another file for this package"
-      linters:
-        - golint
-    - text: "exitAfterDefer:"
-      linters:
-        - gocritic
-    - text: "whyNoLint: include an explanation for nolint directive"
-      linters:
-        - gocritic
-    - text: "go.mongodb.org/mongo-driver/bson/primitive.E"
-      linters:
-        - govet
-    - text: "weak cryptographic primitive"
+    - text: "G114: Use of net/http serve function that has no support for setting timeouts"
       linters:
         - gosec
-    - text: "at least one file in a package should have a package comment"
-      linters:
-        - stylecheck
-    - text: "should have a package comment"
-      linters:
+    - linters:
+        - unparam
         - revive
-    - text: 'Deferring unsafe method "Close" on type "io.ReadCloser"'
-      linters:
+      path: _test\.go$
+      text: "unused-parameter"
+    - linters:
+        - prealloc
+      path: _test\.go$
+      text: "Consider pre-allocating"
+    - linters:
         - gosec
+        - intrange
+      path: _test\.go$
   exclude-use-default: false
