+++ go.mod
@@ -1,8 +1,8 @@
 module github.com/go-pkgz/repeater
 
-go 1.20
+go 1.24
 
-require github.com/stretchr/testify v1.8.2
+require github.com/stretchr/testify v1.10.0
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
