+++ strategy_test.go
