--- v2/go.mod
@@ -1,11 +1,0 @@
-module github.com/go-pkgz/repeater/v2
-
-go 1.24
-
-require github.com/stretchr/testify v1.10.0
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
